generate canopy geometry via script
