+++ package.json
@@ -6,7 +6,8 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "generateCanopy": "ts-node --project tsconfig.script.json src/scripts/generateCanopy.ts"
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.316.0",
@@ -38,6 +39,7 @@
     "recharts": "^2.5.0",
     "splines": "jonathanlurie/splines",
     "three": "^0.151.2",
+    "ts-node": "^10.9.1",
     "use-debounce": "^9.0.4",
     "wavesurfer.js": "^7.0.0-alpha.44"
   },
