+++ tsconfig.script.json
@@ -1,0 +1,6 @@
+{
+  "compilerOptions": {
+    "resolveJsonModule": true,
+    "esModuleInterop": true
+  }
+}
